separate functions list from spans list
