+++ static/app/views/performance/trends/changeExplorerUtils/spansList.tsx
@@ -9,16 +9,13 @@ import LoadingIndicator from 'sentry/components/loadingIndicator';
 import {IconWarning} from 'sentry/icons';
 import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {Organization, Project} from 'sentry/types';
+import {Organization} from 'sentry/types';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import {parsePeriodToHours} from 'sentry/utils/dates';
 import {useDiscoverQuery} from 'sentry/utils/discover/discoverQuery';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import SuspectSpansQuery from 'sentry/utils/performance/suspectSpans/suspectSpansQuery';
 import {SuspectSpan, SuspectSpans} from 'sentry/utils/performance/suspectSpans/types';
-import {EventsResultsDataRow, Sort} from 'sentry/utils/profiling/hooks/types';
-import {useProfileFunctions} from 'sentry/utils/profiling/hooks/useProfileFunctions';
-import {generateProfileFlamechartRouteWithQuery} from 'sentry/utils/profiling/routes';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import useProjects from 'sentry/utils/useProjects';
 import {spanDetailsRouteWithQuery} from 'sentry/views/performance/transactionSummary/transactionSpans/spanDetails/utils';
@@ -61,25 +58,13 @@ export type ChangedSuspectSpan = AveragedSuspectSpan & {
   percentChange: number;
 };
 
-type AveragedSuspectFunction = EventsResultsDataRow<FunctionsField> & {
-  avgSumExclusiveTime: number;
-};
-
-type ChangedSuspectFunction = AveragedSuspectFunction & {
-  avgTimeDifference: number;
-  changeType: string;
-  percentChange: number;
-};
-
-type NumberedListProps = {
+type NumberedSpansListProps = {
   isError: boolean;
   isLoading: boolean;
   limit: number;
   location: Location;
   organization: Organization;
   transactionName: string;
-  functions?: ChangedSuspectFunction[];
-  project?: Project;
   projectID?: string;
   spans?: ChangedSuspectSpan[];
 };
@@ -209,85 +194,6 @@ export function SpansList(props: SpansListProps) {
     ? (totalTransactionsAfter?.data[0]['count()'] as number)
     : 0;
 
-  const functionsSort: Sort<FunctionsField> = {key: 'sum()', order: 'desc'};
-
-  const query = useMemo(() => {
-    const conditions = new MutableSearch('');
-    conditions.setFilterValues('transaction', [transaction.transaction]);
-    return conditions.formatString();
-  }, [transaction.transaction]);
-
-  const beforeFunctionsQuery = useProfileFunctions<FunctionsField>({
-    fields: functionsFields,
-    referrer: 'api.performance.performance-change-explorer',
-    sort: functionsSort,
-    query,
-    limit: 50,
-    datetime: {
-      end: breakpointTime,
-      start: startTime,
-      period: null,
-      utc: null,
-    },
-  });
-
-  const afterFunctionsQuery = useProfileFunctions<FunctionsField>({
-    fields: functionsFields,
-    referrer: 'api.performance.performance-change-explorer',
-    sort: functionsSort,
-    query,
-    limit: 50,
-    datetime: {
-      end: endTime,
-      start: breakpointTime,
-      period: null,
-      utc: null,
-    },
-  });
-
-  // need these averaged fields because comparing total self times may be inaccurate depending on
-  // where the breakpoint is
-  const functionsAveragedAfter = addAvgSumOfFunctions(
-    afterFunctionsQuery.data?.data,
-    transactionCountAfter
-  );
-  const functionsAveragedBefore = addAvgSumOfFunctions(
-    beforeFunctionsQuery.data?.data,
-    transactionCountBefore
-  );
-
-  const addedFunctions = addFunctionChangeFields(
-    findFunctionsNotIn(functionsAveragedAfter, functionsAveragedBefore),
-    true
-  );
-  const removedFunctions = addFunctionChangeFields(
-    findFunctionsNotIn(functionsAveragedBefore, functionsAveragedAfter),
-    false
-  );
-
-  const remainingFunctionsBefore = findFunctionsIn(
-    functionsAveragedBefore,
-    functionsAveragedAfter
-  );
-  const remainingFunctionsAfter = findFunctionsIn(
-    functionsAveragedAfter,
-    functionsAveragedBefore
-  );
-
-  const remainingFunctionsWithChange = addPercentChangeInFunctions(
-    remainingFunctionsBefore,
-    remainingFunctionsAfter
-  );
-
-  const allFunctionsUpdated = remainingFunctionsWithChange
-    ?.concat(addedFunctions ? addedFunctions : [])
-    .concat(removedFunctions ? removedFunctions : []);
-
-  // sorts all functions in descending order of avgTimeDifference (change in avg total self time)
-  const functionList = allFunctionsUpdated?.sort(
-    (a, b) => b.avgTimeDifference - a.avgTimeDifference
-  );
-
   return (
     <SuspectSpansQuery
       location={beforeLocation}
@@ -361,36 +267,28 @@ export function SpansList(props: SpansListProps) {
               );
               // reverse the span list when trendChangeType is improvement so most negative (improved) change is first
               return (
-                <NumberedList
+                <NumberedSpansList
                   spans={
                     trendChangeType === TrendChangeType.REGRESSION
                       ? spanList
                       : spanList?.reverse()
                   }
-                  functions={
-                    trendChangeType === TrendChangeType.REGRESSION
-                      ? functionList
-                      : functionList?.reverse()
-                  }
                   projectID={projectID}
-                  project={projects.find(project => project.id === projectID)}
                   location={location}
                   organization={organization}
                   transactionName={transaction.transaction}
-                  limit={8}
+                  limit={4}
                   isLoading={
                     transactionsLoadingBefore ||
                     transactionsLoadingAfter ||
                     spansLoadingBefore ||
-                    spansLoadingAfter ||
-                    beforeFunctionsQuery.isLoading ||
-                    afterFunctionsQuery.isLoading
+                    spansLoadingAfter
                   }
                   isError={
                     transactionsErrorBefore ||
                     transactionsErrorAfter ||
-                    ((hasSpansErrorBefore || hasSpansErrorAfter) &&
-                      (beforeFunctionsQuery.isError || afterFunctionsQuery.isError))
+                    hasSpansErrorBefore ||
+                    hasSpansErrorAfter
                   }
                 />
               );
@@ -478,34 +376,6 @@ function findSpansIn(
   });
 }
 
-function findFunctionsNotIn(
-  initialFunctions: AveragedSuspectFunction[] | undefined,
-  comparingFunctions: AveragedSuspectFunction[] | undefined
-) {
-  return initialFunctions?.filter(initialValue => {
-    const functionInComparingSet = comparingFunctions?.find(
-      comparingValue =>
-        comparingValue.function === initialValue.function &&
-        comparingValue.package === initialValue.package
-    );
-    return functionInComparingSet === undefined;
-  });
-}
-
-function findFunctionsIn(
-  initialFunctions: AveragedSuspectFunction[] | undefined,
-  comparingFunctions: AveragedSuspectFunction[] | undefined
-) {
-  return initialFunctions?.filter(initialValue => {
-    const functionInComparingSet = comparingFunctions?.find(
-      comparingValue =>
-        comparingValue.function === initialValue.function &&
-        comparingValue.package === initialValue.package
-    );
-    return functionInComparingSet !== undefined;
-  });
-}
-
 /**
  *
  * adds an average of the sumExclusive time so it is more comparable when the breakpoint
@@ -520,25 +390,6 @@ function addAvgSumExclusiveTime(
       ...span,
       avgSumExclusiveTime: span.sumExclusiveTime
         ? span.sumExclusiveTime / transactionCount
-        : 0,
-    };
-  });
-}
-
-/**
- *
- * adds an average of the sum() time so it is more comparable when the breakpoint
- * is not close to the middle of the timeseries
- */
-function addAvgSumOfFunctions(
-  suspectFunctions: EventsResultsDataRow<FunctionsField>[] | undefined,
-  transactionCount: number
-) {
-  return suspectFunctions?.map(susFunc => {
-    return {
-      ...susFunc,
-      avgSumExclusiveTime: (susFunc['sum()'] as number)
-        ? (susFunc['sum()'] as number) / transactionCount
         : 0,
     };
   });
@@ -569,32 +420,6 @@ function addPercentChangeInSpans(
   });
 }
 
-function addPercentChangeInFunctions(
-  before: AveragedSuspectFunction[] | undefined,
-  after: AveragedSuspectFunction[] | undefined
-) {
-  return after?.map(functionAfter => {
-    const functionBefore = before?.find(
-      beforeValue =>
-        functionAfter.function === beforeValue.function &&
-        functionAfter.package === beforeValue.package
-    );
-    const percentageChange =
-      relativeChange(
-        functionBefore?.avgSumExclusiveTime || 0,
-        functionAfter.avgSumExclusiveTime
-      ) * 100;
-    return {
-      ...functionAfter,
-      percentChange: percentageChange,
-      avgTimeDifference:
-        functionAfter.avgSumExclusiveTime - (functionBefore?.avgSumExclusiveTime || 0),
-      changeType:
-        percentageChange < 0 ? SpanChangeType.improved : SpanChangeType.regressed,
-    };
-  });
-}
-
 function addSpanChangeFields(
   spans: AveragedSuspectSpan[] | undefined,
   added: boolean
@@ -619,42 +444,16 @@ function addSpanChangeFields(
   });
 }
 
-function addFunctionChangeFields(
-  functions: AveragedSuspectFunction[] | undefined,
-  added: boolean
-): ChangedSuspectFunction[] | undefined {
-  // percent change is hardcoded to pass the 1% change threshold,
-  // avoid infinite values and reflect correct change type
-  return functions?.map(func => {
-    if (added) {
-      return {
-        ...func,
-        percentChange: 100,
-        avgTimeDifference: func.avgSumExclusiveTime,
-        changeType: SpanChangeType.added,
-      };
-    }
-    return {
-      ...func,
-      percentChange: -100,
-      avgTimeDifference: 0 - func.avgSumExclusiveTime,
-      changeType: SpanChangeType.removed,
-    };
-  });
-}
-
-export function NumberedList(props: NumberedListProps) {
+export function NumberedSpansList(props: NumberedSpansListProps) {
   const {
     spans,
     projectID,
-    project,
     location,
     transactionName,
     organization,
     limit,
     isLoading,
     isError,
-    functions,
   } = props;
 
   if (isLoading) {
@@ -664,36 +463,24 @@ export function NumberedList(props: NumberedListProps) {
   if (isError) {
     return (
       <ErrorWrapper>
-        <IconWarning data-test-id="error-indicator" color="gray200" size="xxl" />
+        <IconWarning data-test-id="error-indicator-spans" color="gray200" size="xxl" />
         <p>{t('There was an issue finding suspect spans for this transaction')}</p>
       </ErrorWrapper>
     );
   }
 
-  if ((spans?.length === 0 || !spans) && (functions?.length === 0 || !functions)) {
+  if (spans?.length === 0 || !spans) {
     return (
       <EmptyStateWarning>
-        <p data-test-id="spans-no-results">{t('No results found for your query')}</p>
+        <p data-test-id="spans-no-results">{t('No results found for suspect spans')}</p>
       </EmptyStateWarning>
     );
-  }
-
-  let spansLimit = limit / 2;
-  let functionsLimit = limit / 2;
-
-  if (spans && (!functions || functions.length === 0)) {
-    spansLimit = limit;
-    functionsLimit = 0;
-  }
-  if (functions && (!spans || spans.length === 0)) {
-    functionsLimit = limit;
-    spansLimit = 0;
   }
 
   // percent change of a span must be more than 1%
   const formattedSpans = spans
     ?.filter(span => (spans.length > 10 ? Math.abs(span.percentChange) >= 1 : true))
-    .slice(0, spansLimit)
+    .slice(0, limit)
     .map((span, index) => {
       const spanDetailsPage = spanDetailsRouteWithQuery({
         orgSlug: organization.slug,
@@ -729,80 +516,32 @@ export function NumberedList(props: NumberedListProps) {
       );
     });
 
-  // percent change of a function must be more than 1%
-  const formattedFunctions = functions
-    ?.filter(func => (functions.length > 10 ? Math.abs(func.percentChange) >= 1 : true))
-    .slice(0, functionsLimit)
-    .map((func, index) => {
-      const profiles = func['examples()'] as string[];
-
-      const functionSummaryView = generateProfileFlamechartRouteWithQuery({
-        orgSlug: organization.slug,
-        projectSlug: project?.slug || '',
-        profileId: profiles[0],
-        query: {
-          frameName: func.function as string,
-          framePackage: func.package as string,
-        },
-      });
-
-      const handleClickAnalytics = () => {
-        trackAnalytics(
-          'performance_views.performance_change_explorer.function_link_clicked',
-          {
-            organization,
-            transaction: transactionName,
-            package: func.package as string,
-            function: func.function as string,
-            profile_id: profiles[0],
-          }
-        );
-      };
-
-      return (
-        <li key={`list-item-${index}`}>
-          <ListItemWrapper data-test-id="list-item">
-            <p style={{marginLeft: space(2)}}>
-              {tct('[changeType] suspect function', {changeType: func.changeType})}
-            </p>
-            <ListLink to={functionSummaryView} onClick={handleClickAnalytics}>
-              {func.function}
-            </ListLink>
-          </ListItemWrapper>
-        </li>
-      );
-    });
-
-  if (formattedSpans?.length === 0 && formattedFunctions?.length === 0) {
+  if (formattedSpans?.length === 0) {
     return (
       <EmptyStateWarning>
-        <p data-test-id="spans-no-changes">
-          {t('No sizable changes in suspect spans and functions')}
-        </p>
+        <p data-test-id="spans-no-changes">{t('No sizable changes in suspect spans')}</p>
       </EmptyStateWarning>
     );
   }
 
   return (
     <div style={{marginTop: space(4)}}>
-      <ol>
-        {formattedSpans}
-        {formattedFunctions}
-      </ol>
+      <h6>{t('Relevant Suspect Spans')}</h6>
+      <ol>{formattedSpans}</ol>
     </div>
   );
 }
 
-const ListLink = styled(Link)`
+export const ListLink = styled(Link)`
   margin-left: ${space(1)};
   ${p => p.theme.overflowEllipsis}
 `;
-const ListItemWrapper = styled('div')`
+export const ListItemWrapper = styled('div')`
   display: flex;
   white-space: nowrap;
 `;
 
-const ErrorWrapper = styled('div')`
+export const ErrorWrapper = styled('div')`
   display: flex;
   margin-top: ${space(4)};
   flex-direction: column;
